+++ torch/csrc/jit/mobile/flatbuffer_loader.h
@@ -1,24 +1,31 @@
 #pragma once
 
+#include <istream>
 #include <string>
 #include <unordered_map>
 #include <vector>
+#include <memory>
 
 #include <ATen/core/ivalue.h>
-#include <caffe2/serialize/inline_container.h>
-#include <torch/csrc/jit/mobile/function.h>
-#include <torch/csrc/jit/mobile/interpreter.h>
+#include <c10/core/Device.h>
+#include <c10/macros/Macros.h>
+#include <c10/util/Optional.h>
 #include <torch/csrc/jit/mobile/module.h>
-#include <torch/csrc/jit/runtime/instruction.h>
-#include <torch/csrc/jit/serialization/mobile_bytecode_generated.h> // NOLINT
-#include <torch/custom_class.h>
 
 /**
  * Defines the public API for loading flatbuffer-serialized mobile modules.
+ * Note that this header must not include or depend on flatbuffer-defined
+ * types, to avoid leaking those details to PyTorch clients.
  */
 
 namespace torch {
 namespace jit {
+
+/// All non-copied data pointers provided to `parse_and_initialize_*` functions
+/// must be aligned to this boundary. Since the Module will point directly into
+/// the data, this alignment is necessary to ensure that certain types/structs
+/// are properly aligned.
+constexpr size_t kFlatbufferDataAlignmentBytes = 16;
 
 /// Maps file names to file contents.
 using ExtraFilesMap = std::unordered_map<std::string, std::string>;
@@ -30,22 +37,9 @@ using ExtraFilesMap = std::unordered_map<std::string, std::string>;
 //    structure
 // 3. Module initialization: Produce mobile::Module out of the structure
 //    produced in 2.
-// Under this context, the structure described in 2. is
-// mobile::serialization::Module
-
-/// DEPRECATED: Use a parse/load function below.
-// Parse a mobile::Module from flatbuffer's in-memory Module representation.
-// The caller is assumed to manage the lifetimes of Module.
-// This function does step 3 described above.
-// If should_copy_tensor_memory is true, then the returned module will NOT
-// have refences to flatbuffer_module, so it can be discarded.
-// If should_copy_tensor_memory is false, then returned module will have
-// tensors that points inside of flatbuffer_module; the caller need to make
-// sure that flatbuffer_module outlives returned Module.
-TORCH_API mobile::Module initialize_mobile_module(
-    mobile::serialization::Module* flatbuffer_module,
-    c10::optional<at::Device> device = c10::nullopt,
-    bool should_copy_tensor_memory = false);
+// Under this context, the structure described in 2. is the flatbuffer-defined
+// type mobile::serialization::Module. However, this step/type is not visible in
+// the public API.
 
 // Parse a mobile::Module from raw bytes.
 //
@@ -59,7 +53,8 @@ TORCH_API mobile::Module initialize_mobile_module(
 //
 // If should_copy_tensor_memory is false, then returned module will have tensors
 // that points inside of `data`; the caller will need to make sure that `data`
-// outlives the returned Module.
+// outlives the returned Module. Also, `data` must be aligned to
+// kFlatbufferDataAlignmentBytes.
 TORCH_API mobile::Module parse_and_initialize_mobile_module(
     void* data,
     size_t size, // of `data`, in bytes.
@@ -71,7 +66,8 @@ TORCH_API mobile::Module parse_and_initialize_mobile_module(
 //
 // This function does steps 2+3 described above.
 //
-// The returned Module holds a reference to `data`.
+// The returned Module holds a reference to `data`, which must be aligned to
+// kFlatbufferDataAlignmentBytes.
 //
 // If you do not want the Module to hold a reference to `data`, see the raw
 // pointer overload of this function.
@@ -107,12 +103,6 @@ TORCH_API mobile::Module load_mobile_module_from_file(
     c10::optional<at::Device> device = c10::nullopt,
     ExtraFilesMap* extra_files = nullptr);
 
-/// DEPRECATED: Use the `extra_files` parameter of one of the parse/load
-/// functions above.
-TORCH_API void parseExtraFiles(
-    mobile::serialization::Module* module,
-    ExtraFilesMap& extra_files);
-
 TORCH_API uint64_t get_bytecode_version(std::istream& in);
 TORCH_API uint64_t get_bytecode_version(const std::string& filename);
 TORCH_API uint64_t get_bytecode_version_from_bytes(char* flatbuffer_content);
@@ -132,98 +122,6 @@ TORCH_API mobile::Module load_mobile_module_from_stream_with_copy(
 // and friends. This is NOT needed if using the other functions
 // in this file directly.
 TORCH_API bool register_flatbuffer_loader();
-
-/// DEPRECATED: Use one of the parse/load functions above.
-class TORCH_API FlatbufferLoader {
- public:
-  FlatbufferLoader();
-
-  typedef IValue (
-      *IValueParser)(FlatbufferLoader&, const mobile::serialization::IValue&);
-  void registerIValueParser(
-      mobile::serialization::IValueUnion ivalue_type,
-      IValueParser parser);
-  mobile::Module parseModule(mobile::serialization::Module* module);
-
-  void extractJitSourceAndConstants(
-      ExtraFilesMap* jit_sources,
-      std::vector<IValue>* constants);
-
-  typedef TypePtr (*TypeResolver)(
-      const std::string& type_str,
-      std::shared_ptr<CompilationUnit> cu);
-
-  void internal_registerTypeResolver(TypeResolver type_resolver);
-
-  IValue& getIValue(uint32_t pos) {
-    TORCH_CHECK(pos < all_ivalues_.size());
-    return all_ivalues_[pos];
-  }
-
-  mobile::Function* getFunction(uint32_t pos) {
-    return all_functions_[pos];
-  }
-
-  ClassTypePtr getType(uint32_t pos) {
-    TORCH_CHECK(pos < all_ivalues_.size());
-    return all_types_[pos];
-  }
-
-  c10::Storage getStorage(uint32_t index);
-  TypePtr getOrCreateTypeAnnotations(const flatbuffers::String* offset);
-  ClassTypePtr getOrCreateClassTypeForObject(
-      const mobile::serialization::Object* object);
-
-  const mobile::serialization::Module* getCurrentFlatbufferInput() {
-    return module_;
-  }
-
-  bool getShouldCopyTensorMemory() {
-    return should_copy_tensor_memory_;
-  }
-
-  void setShouldCopyTensorMemory(bool should_copy_tensor_memory) {
-    should_copy_tensor_memory_ = should_copy_tensor_memory;
-  }
-
-  // Whether or not should load operators in functions.
-  // Not loading operators is useful because if an operator is not found
-  // then we throw exceptions, and sometimes we want to print out
-  // what operators are included before that to debug.
-  void setShouldLoadOperators(bool should_load_operators) {
-    should_load_operators_ = should_load_operators;
-  }
-
-  std::shared_ptr<mobile::CompilationUnit> mcu_;
-  std::shared_ptr<CompilationUnit> cu_;
-
- private:
-  IValue parseIValue(const mobile::serialization::IValue* ivalue);
-  std::unique_ptr<mobile::Function> parseFunction(
-      const mobile::serialization::Function* method);
-  void parseAndPopulate(
-      uint32_t i,
-      const mobile::serialization::IValue* ivalue);
-
-  std::unordered_map<uint32_t, mobile::Function*> all_functions_;
-  std::vector<ClassTypePtr> all_types_;
-  std::unordered_set<uint32_t> initialized_types_;
-  std::unordered_map<const flatbuffers::String*, TypePtr> type_annotations_;
-  std::vector<bool> storage_loaded_;
-  std::vector<c10::Storage> storages_;
-  std::vector<IValue> all_ivalues_;
-  std::array<
-      IValueParser,
-      static_cast<uint8_t>(mobile::serialization::IValueUnion::MAX) + 1>
-      ivalue_parsers_;
-  TypeResolver type_resolver_ = nullptr;
-  mobile::serialization::Module* module_ = nullptr;
-  bool module_parsed_ = false;
-  bool should_copy_tensor_memory_ = false;
-  bool should_load_operators_ = true;
-  // 0 -> mobile_ivalue_size_ elements are from the mobile module.
-  uint32_t mobile_ivalue_size_ = 0;
-};
 
 } // namespace jit
 } // namespace torch
